claude_demo/requirements.txt
torch
torchvision
transformers
opencv-python
pillow
numpy
pandas
sentence-transformers
faiss-cpu
gradio
tqdm
accelerate
datasets